Implemented dynamic CHS fitting in BIOS
Instead of using something hardcoded, the BIOS now finds a fit for the card,
given the size it reports.

Constraints:
 - Between 2 and 255 heads
 - Prefer more sectors per track than more heads (formatting is quicker)

On some devices, if a perfect match is not found, the few last sectors
will not be accessible.

If the card is larger than what 1024 cylinders, 255 heads and 63 sectors
can address, a size of 1024/255/63 is returned.