Add a convenience function to clock, deselect, and re-clock
Send 24 clock pulses, deselects the card, then sends another 24 pulses.

The ideal for the first 16 pulses comes from reading (and discarding) the
two-byte CRC after reading a sector. The extra byte before deselect,
as well as the extra 3 bytes after it are there in the hope it helps
with some cards. (Some code here and there on the net suggest this
is useful, and it seems to help on one of my cards which generated
more or less random errors before doing this)