Support for co-existance with another hard drive such as JR-IDE
SD-Cart JR now detects if another hard drive BIOS is present, and
does the necessary arrangements to become the last drive in the system.

A new boot menu allows this behavior to be modified. The SD-Cart JR
can become the first drive, and the other hard drive second drive.
Installing the SD-Cart JR as a sole drive (equivalent to previous
releases), or not installing it at all is also possible.