Make fdisk multiple disk support work
For instance, when SD-Cart JR is installed as "last drive", the count in DL
returned for drive 80h is correct (it counts original drives, for
instance the jr-IDE, plus one for the SD-Cart JR drive). However,
the count returned for drive 81h is still 1. This value needs to be
adjusted, otherwise fdisk does not offer the 5th option to select which
fixed disk to work on. (DOS however did not care)